doc: Remove loro abstraction
Creating and subscribing to a loro doc is quite straight forward. I
think having an abstraction between the `Document` and loro makes the
code unnecessarily more complicated. Also if we decide to run the CRDT
in a distinct thread we can use the `Document` object directly since
it's Send+Sync.